chore: refactor LeadOptionBottomSheet import path and create new component for lead options
